+++ tasm-lib/src/snippet.rs
@@ -5,6 +5,7 @@ use std::fmt::Display;
 use std::rc::Rc;
 
 use anyhow::Result;
+use itertools::Itertools;
 use triton_vm::instruction::LabelledInstruction;
 use triton_vm::parser::{to_labelled_instructions, tokenize};
 use triton_vm::{triton_asm, NonDeterminism, Program};
@@ -142,7 +143,7 @@ pub trait DeprecatedSnippet {
 
     fn link_for_isolated_run(
         &self,
-        words_statically_allocated: Option<usize>,
+        words_statically_allocated: Option<u32>,
     ) -> Vec<LabelledInstruction> {
         let mut snippet_state = if let Some(number_of_words) = words_statically_allocated {
             Library::with_preallocated_memory(number_of_words)
@@ -178,7 +179,7 @@ pub trait DeprecatedSnippet {
         std_in: Vec<BFieldElement>,
         secret_in: Vec<BFieldElement>,
         memory: HashMap<BFieldElement, BFieldElement>,
-        words_allocated: Option<usize>,
+        words_allocated: Option<u32>,
     ) -> Result<VmOutputState> {
         let expected_length_prior: usize = self.input_types().iter().map(|x| x.stack_size()).sum();
         let expected_length_after: usize = self.output_types().iter().map(|x| x.stack_size()).sum();
@@ -219,7 +220,7 @@ pub trait DeprecatedSnippet {
         std_in: Vec<BFieldElement>,
         secret_in: Vec<BFieldElement>,
         memory: &mut HashMap<BFieldElement, BFieldElement>,
-        words_statically_allocated: Option<usize>,
+        words_statically_allocated: Option<u32>,
     ) -> Result<ExecutionResult> {
         let expected_length_prior: usize = self.input_types().iter().map(|x| x.stack_size()).sum();
         let expected_length_after: usize = self.output_types().iter().map(|x| x.stack_size()).sum();
